requirements.txt
eventlet>=0.33.0
python-dotenv
feedparser
openai
flask
apscheduler
Flask-SocketIO
requests